# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Gaze Engine API
  version: 3.0.0
  description: Privacy-first local video search engine API
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "48100"
security:
  - bearerAuth: []
paths:
...
  /videos:
    get:
      operationId: listVideos
      summary: List videos with pagination and filtering
      parameters:
        - name: library_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideosResponse"
  /videos/{id}:
    get:
      operationId: getVideo
      summary: Get video details and indexing status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /videos/{id}/frames:
    get:
      operationId: listVideoFrames
      summary: List sampled thumbnail frames for a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 15
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: List of frames
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FramesResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Video:
      type: object
      required:
        - video_id
        - library_id
        - path
        - filename
        - status
        - created_at_ms
      properties:
        video_id:
          type: string
        library_id:
          type: string
        path:
          type: string
        filename:
          type: string
        file_size:
          type: integer
        duration_ms:
          type: integer
        width:
          type: integer
        height:
          type: integer
        status:
          type: string
          enum: [QUEUED, EXTRACTING_AUDIO, TRANSCRIBING, EXTRACTING_FRAMES, EMBEDDING, DETECTING, DONE, FAILED, CANCELLED]
        progress:
          type: number
          minimum: 0
          maximum: 1
        error_message:
          type: string
        thumbnail_path:
          type: string
        created_at_ms:
          type: integer
        indexed_at_ms:
          type: integer
    Frame:
      type: object
      required:
        - frame_index
        - timestamp_ms
        - thumbnail_path
      properties:
        frame_index:
          type: integer
        timestamp_ms:
          type: integer
        thumbnail_path:
          type: string
    FramesResponse:
      type: object
      required:
        - frames
        - total
      properties:
        frames:
          type: array
          items:
            $ref: "#/components/schemas/Frame"
        total:
          type: integer
    VideosResponse:
      type: object
      required:
        - videos
        - total
      properties:
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        total:
          type: integer
...